+++ app/Http/Controllers/SectionController.php
@@ -131,10 +131,14 @@ class SectionController extends Controller
         $AlertType = "";
         $AlertMsg = "";
         try {
-            Section::whereIn('id', $request->section_ids)->delete();
-
-            $AlertType = "success";
-            $AlertMsg = "Selected data deleted";
+            if ($request->group_ids != null) {
+                Section::whereIn('id', $request->section_ids)->delete();
+                $AlertType = "success";
+                $AlertMsg = "Selected data deleted";
+            } else {
+                $AlertType = "warning";
+                $AlertMsg = "Please select atleast one row.";
+            }
         } catch (\Illuminate\Database\QueryException $ex) {
             if ($ex->getCode() == 23000) {
                 $AlertType = "danger";
